clean: Clean up messages to user
The logging of the removed folders did not account for folders that
were fell within the desired aged to delete but could not be removed
because they are still in use. Thus they initial message about the
number of folders found flew in face of the number actually deleted.

Added a count of rejected deletions on the assumption it was because
they are in use still.

modified the messaging to account for the rejected deletions

removed the number deleted logging from the index as no longer being
informative