clean: Clean up messages to user (#146)
## Problem

The logging of the removed folders did not account for folders that fell
within the desired aged to delete but could not be removed because they
are still in use. Thus they initial message about the number of folders
found did not agree with the number actually deleted.

## Changes

- Added a count of rejected deletions on the assumption it was because
they are in use still.
- modified the messaging to account for the rejected deletions
- removed the number deleted logging from the index as no longer being
informative

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Enhanced logging for folder deletion outcomes in the Sonar temporary
folder removal process, providing clearer feedback on operation success.

- **Bug Fixes**
- Improved feedback mechanism for folder deletion, addressing visibility
of successful and failed deletions.

- **Documentation**
- Added comments to clarify the purpose of functions related to
temporary folders created by SonarLint.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>